feat: show ev direction on frontend
